Project-7/app/requirements.txt
fastapi
requests
Jinja2
python-multipart
uvicorn
python-dotenv
numpy
scikit-learn
scipy
